+++ src/Libraries/SmartStore.Core/Domain/Customers/PrivacySettings.cs
@@ -35,8 +35,13 @@ namespace SmartStore.Core.Domain.Customers
 		public bool DisplayGdprConsentOnForms { get; set; }
 
 		/// <summary>
-		/// Gets or sets a value indicating whether to display a checkbox to the customer where he can agree to privacy terms
+		/// Gets or sets a value indicating whether the full name field is required on contact us requests
 		/// </summary>
 		public bool FullNameOnContactUsRequired { get; set; }
+
+		/// <summary>
+		/// Gets or sets a value indicating whether the full name field is required on product requests
+		/// </summary>
+		public bool FullNameOnProductRequestRequired { get; set; }
 	}
 }